fix: use int64 for SendLocation MessageTimestamp
Review feedback on #44 (which also applies here): ms timestamps exceed
int32 range. Widen SendLocationResponse.MessageTimestamp to int64 and
drop the int() cast in the handler.

+++ server/dto/message.go
@@ -24,7 +24,7 @@ type SendLocationResponse struct {
 	Key              MessageResponseKey `json:"key"`
 	Status           string             `json:"status"`
 	MessageType      string             `json:"messageType"`
-	MessageTimestamp int                `json:"messageTimestamp"`
+	MessageTimestamp int64              `json:"messageTimestamp"`
 	InstanceId       string             `json:"instanceId"`
 }
